reorganizing project files
