+++ app.py
@@ -1,0 +1,7 @@
+import os
+from src import app
+
+
+if __name__ == '__main__':
+  port = int(os.environ.get('PORT', 5000))  # default to 5000
+  app.run(host='0.0.0.0', port=port)
