+++ src/configs/dev.py
@@ -1,0 +1,5 @@
+import os
+
+DEBUG = True
+SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
+SQLALCHEMY_TRACK_MODIFICATIONS = False
