+++ src/configs/test.py
@@ -1,0 +1,12 @@
+import os
+import subprocess
+
+DEBUG = True
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+if 'TEST_DB_URL' in os.environ:
+  SQLALCHEMY_DATABASE_URI = os.environ.get('TEST_DB_URL')
+else:
+  SQLALCHEMY_DATABASE_URI = subprocess.check_output(
+    "heroku config --app quokka-api-test-db | grep DATABASE_URL | awk '{print $2}'",
+    shell=True).decode('utf-8')
