+++ src/helpers/definitions.py
@@ -1,0 +1,7 @@
+import os
+
+basedir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..'))
+
+source_dir = basedir + '/src'
+
+configs_dir = source_dir + '/configs'
