+++ src/helpers/env.py
@@ -1,0 +1,22 @@
+import os
+
+environment = os.environ.get('ENV')
+assert environment
+
+ENV = environment.lower()
+
+
+def env():
+  return ENV
+
+
+def is_test():
+  return ENV == 'test'
+
+
+def is_dev():
+  return ENV == 'dev'
+
+
+def is_prod():
+  return ENV == 'prod'
