+++ src/helpers/error_codes.py
@@ -1,0 +1,9 @@
+"""
+Custom Error Codes for API responses
+
+See https://github.com/quokkateam/quokka/blob/3af7fec346c63dec27c11a8c302de63841c3182e/src/utils/StatusCodes.js
+"""
+
+# User Validation
+INVALID_EMAIL_DOMAIN = 1200
+MALFORMED_EMAIL = 1201
