+++ src/integrations/email.py
@@ -1,2 +1,2 @@
 def send_verification_email(user):
-    pass
+  pass
