--- src/logger.py
@@ -1,3 +1,0 @@
-import logging
-
-logger = logging.getLogger('quokka')
