--- src/manage.py
@@ -1,10 +1,0 @@
-from app import create_app
-from flask_migrate import MigrateCommand
-from flask_script import Manager
-
-manager = Manager(create_app())
-manager.add_option("-c", "--config", dest="config_module", required=False)
-manager.add_command('db', MigrateCommand)
-
-if __name__ == '__main__':
-  manager.run()
