+++ src/snapshots/snap_prod_config_snapshot_test.py
@@ -7,8 +7,4 @@ from snapshottest import Snapshot
 
 snapshots = Snapshot()
 
-snapshots['test_prod_config 1'] = '''{
-  "DEBUG": false,
-  "SQLALCHEMY_DATABASE_URI": "the-db-url",
-  "SQLALCHEMY_TRACK_MODIFICATIONS": false
-}'''
+snapshots['test_prod_config 1'] = '{}'
